UnfoldMakie.cart3d_to_spherical — Method

cart3d_to_spherical(x, y, z)

Convert x, y, z electrode positions on a scalp to spherical coordinate representation.

Return Value: Matrix.

UnfoldMakie.config_kwargs! — Method

config_kwargs!(cfg::PlotConfig; kwargs...)

Takes NamedTuple of Key => NamedTuple as kwargs and merges the fields with the defaults.

UnfoldMakie.data_binning — Method

data_binning(df, bin_width; col_y = :erp, fun = mean, grouping = [])

Group DataFrame according to topoplot coordinates and apply aggregation function.

Arguments:

df::AbstractTable
Requires columns :cont_cuts, col_y (default :erp), and all columns in grouping (col_coord, row_coord, label);

col_y = :erp
The column to combine over (with fun);

fun = mean()
Function to combine.

grouping = []
Vector of symbols or strings, columns to group by the data before aggregation. Values of nothing are ignored.

Return Value: DataFrame.

UnfoldMakie.eeg_array_to_dataframe — Method

eeg_array_to_dataframe(data::AbstractMatrix, label_aliases::AbstractVector)
eeg_array_to_dataframe(data::AbstractVector, label_aliases::AbstractVector)
eeg_array_to_dataframe(data::Union{AbstractMatrix, AbstractVector{<:Number}})

Helper function converting an array (Matrix or Vector) to a tidy DataFrame with columns :estimate, :time and :label (with aliases :color, :group, :channel).

Format of Arrays:
- times x condition for plot_erp.
- channels x time for plot_butterfly, plot_topoplotseries.
- channels for plot_topoplot.
Return Value: DataFrame.

UnfoldMakie.eeg_topoplot_series — Method

eeg_topoplot_series(data::DataFrame,
    f,
    data::DataFrame;
    bin_width,
    bin_num,
    y = :erp,
    label = :label,
    col = :time,
    row = nothing,
    col_labels = false,
    row_labels = false,
    rasterize_heatmaps = true,
    combinefun = mean,
    xlim_topo,
    ylim_topo,
    topoplot_attributes...,
)
eeg_topoplot_series!(fig, data::DataFrame, bin_width; kwargs..)

Plot a series of topoplots. The function takes the combinefun = mean over the :time column of data in bin_width steps.

f
Figure object.

data::DataFrame
Needs the columns :time and y(=:erp), and label(=:label).
If data is a matrix, it is automatically cast to a dataframe, time bins are in samples, labels are string.(1:size(data,1)).

col, row = :time
Specify the field to be divided into columns and rows. The default is col = :time to split by the time field and row = nothing.
Useful to split by a condition, e.g. ...(..., col = :time, row = :condition) would result in multiple (as many as different values in df.condition) rows of topoplot series.

row_labels, col_labels = false
Indicate whether there should be labels in the plots in the first column to indicate the row value and in the last row to indicate the time (typically timerange).

combinefun::Function = mean
Specify how the samples within bin_width are summarised.
Example functions: mean, median, std.

Return Value: Tuple{Figure, Vector{Any}}.

UnfoldMakie.plot_butterfly — Method

plot_butterfly(plot_data::Union{DataFrame, AbstractMatrix}; kwargs...)
plot_butterfly(times::Vector, plot_data::Union{DataFrame, AbstractMatrix}; kwargs...)
plot_butterfly!(f::Union{GridPosition, GridLayout, Figure}, plot_data::Union{DataFrame, AbstractMatrix}; kwargs...)

Plot a Butterfly plot.

Arguments

f::Union{GridPosition, GridLayout, Figure}
Figure, GridLayout, or GridPosition to draw the plot.

data::Union{DataFrame, AbstractMatrix}
Data for the ERP plot visualization.

kwargs...
Additional styling behavior.
Often used as: plot_butterfly(df; visual = (; colormap = :romaO)).

Keyword arguments (kwargs)

positions::Array = []
Adds a topoplot as an inset legend to the provided channel positions. Must be the same length as plot_data. To change the colors of the channel lines use the topoposition_to_color function.

topolegend::Bool = true
Show an inlay topoplot with corresponding electrodes. Requires positions.

topomarkersize::Real = 10
Change the size of the electrode markers in topoplot.

topowidth::Real = 0.25
Change the width of inlay topoplot.

topoheight::Real = 0.25
Change the height of inlay topoplot.

topopositions_to_color::x -> pos_to_color_RomaO(x)
Change the line colors.

Return Value: Figure displaying Butterfly plot.

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = (xlabel = "Time [s]", ylabel = "Voltage [µV]", yticklabelsize = 14, xtickformat = "{:.1f}") - use kwargs... of Makie.Axis

layout = (show_legend = false, use_colorbar = true, use_legend = true, hidespines = (:r, :t), hidedecorations = Dict{Symbol, Bool}(:label => 0, :ticks => 0, :ticklabels => 0))

mapping = (x = (:time,), y = (:estimate, :yhat, :y), color = (:channel, :channels, :trial, :trials), positions = (:pos, :positions, :position, :topo_positions, :x, nothing), labels = (:labels, :label, :topoLabels, :sensor, nothing), group = (:channel,))

visual = (colormap = nothing, color = nothing) - use kwargs... of Makie.lines

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center, framevisible = false) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false, label = "", flipaxis = true, labelrotation = -1.5707963267948966) - use kwargs... of AlgebraOfGraphics.colorbar!

see also plot_erp

UnfoldMakie.plot_channelimage — Method

plot_channelimage!(f::Union{GridPosition, GridLayout, Figure}, data::Union{DataFrame,AbstractMatrix}, positions::Vector{Point{2,Float32}}, ch_names::Vector{String}; kwargs...)
plot_channelimage(data::Union{DataFrame, AbstractMatrix}, positions::Vector{Point{2,Float32}}, ch_names::Vector{String}; kwargs...)

Plot a Channel image

Arguments

f::Union{GridPosition, GridLayout, Figure}
Figure, GridLayout, or GridPosition to draw the plot.

data::Union{DataFrame, AbstractMatrix}
DataFrame or Matrix with data.
Data should has a format of 1 row - 1 channel.

positions::Vector{Point{2,Float32}}
A vector with EEG layout coordinates.

ch_names::Vector{String}
Vector with channel names.

times::Vector = range(-0.3, 1.2, length = size(data, 2))
Time range on x-axis.

sorting_variables::Vector = [:y, :x]
Method to sort channels on y-axis.
For instance, you can sort by channel positions on the scalp (x, y) or channel name.

sorting_reverse::Vector = [:false, :false]
Should sorting variables be reversed or not?

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = (xlabel = "Time [s]", ylabel = "Channels", yticklabelsize = 14) - use kwargs... of Makie.Axis

layout = (show_legend = true, use_colorbar = false)

mapping = (x = (:time,), y = (:estimate, :yhat, :y))

visual = (colormap = Makie.Reverse{String}("RdBu"),) - use kwargs... of Makie.heatmap

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false, label = "Voltage [µV]", labelrotation = -1.5707963267948966) - use kwargs... of Makie.Colorbar

Return Value: Figure displaying the Channel image.

UnfoldMakie.plot_circular_topoplots — Method

plot_circular_topoplots!(f, data::DataFrame; kwargs...)
plot_circular_topoplots(data::DataFrame; kwargs...)

Plot a circular EEG topoplot.

Arguments

f::Union{GridPosition, GridLayout, Figure}
Figure, GridLayout, or GridPosition to draw the plot.

data::DataFrame
DataFrame with data keys (columns :y, :yhat, :estimate), and :position (columns :pos, :position, :positions).

Keyword arguments (kwargs)

predictor::Vector{Any} = :predictor
The circular predictor value, defines position of topoplot across the circle. Mapped around predictor_bounds.

predictor_bounds::Vector{Int64} = [0, 360]
The bounds of the predictor. Relevant for the axis labels.

positions::Vector{Point{2, Float32}} = nothing
Positions of the plot_topoplot.

center_label::String = ""
The text in the center of the cricle.

plot_radius::String = 0.8
The radius of the circular topoplot series plot calucalted by formula: radius = (minwidth * plot_radius) / 2.

labels::Vector{String} = nothing
Labels for the plot_topoplots.

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = (xlabel = "", aspect = Makie.DataAspect(), label = "") - use kwargs... of Makie.Axis

layout = (show_legend = false, use_colorbar = true, hidespines = (), hidedecorations = Dict{Symbol, Bool}(:label => 0))

mapping = (x = nothing, y = (:estimate, :yhat, :y), positions = (:pos, :positions, :position, nothing), labels = (:labels, :label, :sensor, nothing))

visual = (colormap = Makie.Reverse{Symbol}(:RdBu), contours = (color = :white, linewidth = 2), enlarge = 1, label_scatter = true, label_text = true, bounding_geometry = GeometryBasics.Circle) - use kwargs... of Topoplot.eeg_topoplot

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false, flipaxis = true, labelrotation = -1.5707963267948966, label = "Voltage [µV]", colormap = Makie.Reverse{Symbol}(:RdBu)) - use kwargs... of Makie.Colorbar

Return Value: Figure displaying the Circular topoplot series.

UnfoldMakie.plot_designmatrix — Method

plot_designmatrix!(f::Union{GridPosition, GridLayout, Figure}, data::Unfold.DesignMatrix; kwargs...)
plot_designmatrix(data::Unfold.DesignMatrix; kwargs...)

Plot a designmatrix.

Arguments

f::Union{GridPosition, GridLayout, Figure}
Figure, GridLayout, or GridPosition to draw the plot.

data::Unfold.DesignMatrix
Data for the plot visualization.

Keyword arguments (kwargs)

standardize_data::Bool = true
Indicates whether the data is standardized by pointwise division of the data with its sampled standard deviation.

sort_data::Bool = true
Indicates whether the data is sorted. It uses sortslices() of Base Julia.

xticks::Num = nothing
Returns the number of labels on the x axis.

xticks = 0: no labels are placed.

xticks = 1: first possible label is placed.

xticks = 2: first and last possible labels are placed.

2 < xticks < number of labels: equally distribute the labels.

xticks ≥ number of labels: all labels are placed.

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = (xlabel = "Conditions", ylabel = "Trials", xticklabelrotation = 0.39) - use kwargs... of Makie.Axis

layout = (show_legend = true, use_colorbar = true)

mapping = (x = (:time,), y = (:estimate, :yhat, :y))

visual = (colormap = :roma,) - use kwargs... of Makie.heatmap

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false, flipaxis = true, labelrotation = -1.5707963267948966, label = "") - use kwargs... of Makie.Colorbar

Return Value: Figure displaying the Design matrix.

UnfoldMakie.plot_erp — Method

plot_erp!(f::Union{GridPosition, GridLayout, Figure}, plot_data::Union{DataFrame, AbstractMatrix, AbstractVector{<:Number}}; kwargs...)
plot_erp(times, plot_data::Union{DataFrame, AbstractMatrix, AbstractVector{<:Number}}; kwargs...)

Plot an ERP plot.

Arguments

f::Union{GridPosition, GridLayout, Figure}
Figure, GridLayout, or GridPosition to draw the plot.

data::Union{Union{DataFrame, AbstractMatrix, AbstractVector{<:Number}, Vector{Float32}}
Data for the ERP plot visualization.

kwargs...
Additional styling behavior.
Often used as: plot_erp(df; mapping = (; color = :coefname, col = :conditionA)).

Keyword arguments (kwargs)

stderror::Bool = false
Add an error ribbon, with lower and upper limits based on the :stderror column.

significance::DataFrame = nothing
Show significant time periods as horizontal bars.
Example: DataFrame(from = [0.1, 0.3], to = [0.5, 0.7], coefname = ["(Intercept)", "condition: face"]).
If coefname is not specified, the significance lines will be black.

layout.use_colorbar = true
Enable or disable colorbar.

layout.use_legend = true
Enable or disable legend.

layout.show_legend = true
Enable or disable legend and colorbar.

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = (xlabel = "Time [s]", ylabel = "Voltage [µV]", yticklabelsize = 14, xtickformat = "{:.1f}") - use kwargs... of Makie.Axis

layout = (show_legend = true, use_colorbar = true, use_legend = true, hidespines = (:r, :t), hidedecorations = Dict{Symbol, Bool}(:grid => 1, :label => 0, :ticks => 0, :ticklabels => 0))

mapping = (x = (:time,), y = (:estimate, :yhat, :y), color = (:color, :coefname, nothing))

visual = (colormap = :roma,) - use kwargs... of Makie.lines

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center, framevisible = false) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false, label = "", flipaxis = true, labelrotation = -1.5707963267948966) - use kwargs... of AlgebraOfGraphics.colorbar!

Return Value: Figure displaying the ERP plot.

UnfoldMakie.plot_erpgrid — Method

plot_erpgrid(data::Union{Matrix{<:Real}, DataFrame}, positions::Vector; kwargs...)
plot_erpgrid!(f::Union{GridPosition, GridLayout, Figure}, data::Union{Matrix{<:Real}, DataFrame}, positions::Vector, ch_names::Vector{String}; kwargs...)

Plot an ERP image.

Arguments

f::Union{GridPosition, GridLayout, Figure}
Figure, GridLayout, or GridPosition to draw the plot.

data::Union{Matrix{<:Real}, DataFrame}
Data for the plot visualization.
Data should has a format of 1 row - 1 channel.

positions::Vector{Point{2,Float}}
Electrode positions.

ch_names::Vector{String}
Vector with channel names.

Keyword arguments (kwargs)

drawlabels::Bool = false
Draw channels labels over each waveform.

times::Vector = 1:size(data, 2)
Vector of size().

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = (xlabel = "Time [s]", ylabel = "Voltage [µV]", xlim = [-0.04, 1.0], ylim = [-0.04, 1.0]) - use kwargs... of Makie.Axis

layout = (show_legend = true, use_colorbar = false)

mapping = (x = (:time,), y = (:estimate, :yhat, :y))

visual = (colormap = :roma,) - use kwargs... of Makie.lines

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false) - use kwargs... of Makie.Colorbar

Return Value: Figure displaying ERP grid.

UnfoldMakie.plot_erpimage — Method

plot_erpimage!(f::Union{GridPosition, GridLayout, Figure}, data::AbstractMatrix{Float64}; kwargs...)
plot_erpimage!(f::Union{GridPosition, GridLayout, Figure}, data::Observable{<:AbstractMatrix}; kwargs...)
plot_erpimage!(f::Union{GridPosition, GridLayout, Figure}, times::Observable{<:AbstractVector}, data::Observable{<:AbstractMatrix{<:Real}}; kwargs...)

plot_erpimage(times::AbstractVector, data::Union{<:Observable{Matrix{<:Real}}, Matrix{<:Real}}; kwargs...)
plot_erpimage(data::Matrix{Float64}; kwargs...)

Plot an ERP image.

Arguments:

f::Union{GridPosition, GridLayout, Figure}
Figure, GridLayout, or GridPosition to draw the plot.

data::Union{DataFrame, Vector{Float32}}
Data for the plot visualization.

Keyword arguments (kwargs)

erpblur::Number = 10
Number indicating how much blur is applied to the image.
Gaussian blur of the ImageFiltering module is used.
Non-Positive values deactivate the blur.

sortvalues::Vector{Int64} = false
Parameter over which plot will be sorted. Using sortperm() of Base Julia.
sortperm() computes a permutation of the array's indices that puts the array in sorted order.

sortindex::Vector{Int64} = nothing
Sorting over index values.

meanplot::bool = false
Add a line plot below the ERP image, showing the mean of the data.

show_sortval::bool = false
Add a plot on the right from ERP image, showing the distribution of the sorting data.

sortval_xlabel::String = "Sorting variable"
If show_sortval = true controls xlabel.

axis.ylabel::String = "Trials"
If sortvalues = true the default text will change to "Sorted trials", but it could be changed to any values specified manually.

meanplot_axis::NamedTuple = (;)
Here you can flexibly change configurations of meanplot.
To see all options just type ?Axis in REPL.

sortplot_axis::NamedTuple = (;)
Here you can flexibly change configurations of meanplot.
To see all options just type ?Axis in REPL.

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = (xlabel = "Time [s]", ylabel = "Trials") - use kwargs... of Makie.Axis

layout = (show_legend = false, use_colorbar = true)

mapping = (x = (:time,), y = (:estimate, :yhat, :y))

visual = (colormap = Makie.Reverse{String}("RdBu"),) - use kwargs... of Makie.heatmap

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false, label = "Voltage [µV]", labelrotation = -1.5707963267948966) - use kwargs... of Makie.Colorbar

Return Value: Figure displaying the ERP image.

UnfoldMakie.plot_parallelcoordinates — Method

plot_parallelcoordinates(data::Union{DataFrame, AbstractMatrix}; kwargs...)
plot_parallelcoordinates!(f::Union{GridPosition, GridLayout, Figure}, data::Union{DataFrame, AbstractMatrix}; kwargs)

Plot a PCP (parallel coordinates plot).
Dimensions: conditions, channels, time, trials.

Arguments:

f::Union{GridPosition, GridLayout, Figure} Figure, GridLayout, or GridPosition to draw the plot.

data::Union{DataFrame, AbstractMatrix}
Data for the plot visualization.

Keyword arguments (kwargs)

normalize::Symbol = nothing
If :minmax, normalize each axis to their respective min-max range.

ax_labels::Vector{String} = nothing
Specify axis labels.
Should be a vector of labels with length equal to the number of unique mapping.x values.
Example: ax_labels = ["Fz", "Cz", "O1", "O2"].

ax_ticklabels::Symbol = :outmost
Specify tick labels on axis.

:all - show all labels on all axes.

:left - show all labels on the left axis, but only min and max on others.

:outmost - show labels on min and max of all other axes.

:none - remove all labels.

bend::Bool = false
Change straight lines between the axes to curved ("bent") lines using spline interpolation.
Note: While this makes the plot look cool, it is not generally recommended to bent the lines, as interpretation suffers, and the resulting visualizations can be potentially missleading.

visual.alpha::Number = 0.5
Change of line transparency.

Defining the axes

mapping.x = :channel, mapping.y = :estimate.
Overwrite what should be on the x and the y axes.

mapping.color = :colorcolumn
Split conditions by color. The default color is :black.

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = (xlabel = "Channels", ylabel = "Time", title = "") - use kwargs... of Makie.Axis

layout = (show_legend = true, use_colorbar = false)

mapping = (x = :channel, y = (:estimate, :yhat, :y))

visual = (colormap = ColorTypes.RGBA{Float32}[RGBA{Float32}(0.0f0,0.44705883f0,0.69803923f0,1.0f0), RGBA{Float32}(0.9019608f0,0.62352943f0,0.0f0,1.0f0), RGBA{Float32}(0.0f0,0.61960787f0,0.4509804f0,1.0f0), RGBA{Float32}(0.8f0,0.4745098f0,0.654902f0,1.0f0), RGBA{Float32}(0.3372549f0,0.7058824f0,0.9137255f0,1.0f0), RGBA{Float32}(0.8352941f0,0.36862746f0,0.0f0,1.0f0), RGBA{Float32}(0.9411765f0,0.89411765f0,0.25882354f0,1.0f0)], color = :black, alpha = 0.3) - use kwargs... of Makie.lines

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center, title = "Conditions", merge = true, framevisible = false) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false) - use kwargs... of Makie.Colorbar

Return Value: Figure displaying the Parallel coordinates plot.

UnfoldMakie.plot_splines — Method

plot_splines(m::UnfoldModel; kwargs...)
plot_splines!(f::Union{GridPosition, GridLayout, Figure}, m::UnfoldModel; kwargs...)

Visualization of spline terms in an UnfoldModel. Two subplots are generated for each spline term:
1) the basis function of the spline. 2) the density of the underlying covariate.
Multiple spline terms are arranged across columns. Dashed lines indicate spline knots.

Arguments:

f::Union{GridPosition, GridLayout, Figure} Figure, GridLayout, or GridPosition to draw the plot.

m::UnfoldModel
UnfoldModel with splines.

spline_axis::NamedTuple = (;)
Here you can flexibly change configurations of spline subplots.
To see all options just type ?Axis in REPL.

density_axis::NamedTuple = (;)
Here you can flexibly change configurations of density subplots.
To see all options just type ?Axis in REPL.

superlabel_axis::NamedTuple = (;)
Here you can flexibly change configurations of the Label on the top of the plot.
To see all options just type ?Label in REPL.

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = NamedTuple() - use kwargs... of Makie.Axis

layout = (show_legend = true, use_colorbar = false)

mapping = (x = (:time,), y = (:estimate, :yhat, :y))

visual = (colormap = :viridis,) - use kwargs... of Makie.series

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center, title = "Splines", framevisible = false) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false) - use kwargs... of Makie.Colorbar

Return Value: Figure with splines and their density for basis functions.

UnfoldMakie.plot_topoplot — Method

plot_topoplot!(f::Union{GridPosition, GridLayout, Figure}, data::Union{<:AbstractDataFrame,<:AbstractVector}; positions::Vector, labels = nothing, kwargs...)
plot_topoplot(data::Union{<:AbstractDataFrame,<:AbstractVector}; position::Vector, labels = nothing, kwargs...)

Plot a topoplot.

Arguments

f::Union{GridPosition, GridLayout, Figure}
Figure, GridLayout, or GridPosition to draw the plot.

data::Union{DataFrame, Vector{Float32}}
Data for the plot visualization.

positions::Vector{Point{2, Float32}}
Positions used if data is not a DataFrame. Positions are generated from labels if positions = nothing.

labels::Vector{String} = nothing
Labels used if data is not a DataFrame.

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = (xlabel = "", aspect = Makie.DataAspect()) - use kwargs... of Makie.Axis

layout = (show_legend = true, use_colorbar = true, hidespines = (), hidedecorations = Dict{Symbol, Bool}(:label => 0))

mapping = (x = nothing, y = (:estimate, :yhat, :y), positions = (:pos, :positions, :position, nothing), labels = (:labels, :label, :sensor, nothing))

visual = (colormap = Makie.Reverse{Symbol}(:RdBu), contours = (color = :white, linewidth = 2), enlarge = 1, label_scatter = true, label_text = true, bounding_geometry = GeometryBasics.Circle) - use kwargs... of Topoplot.eeg_topoplot

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false, flipaxis = true, labelrotation = -1.5707963267948966, label = "Voltage [µV]") - use kwargs... of Makie.Colorbar

Return Value: Figure displaying the Topoplot.

UnfoldMakie.plot_topoplotseries — Method

plot_topoplotseries(f::Union{GridPosition, GridLayout, Figure}, data::Union{<:Observable{<:DataFrame},DataFrame}; kwargs...)
plot_topoplotseries!(data::Union{<:Observable{<:DataFrame},DataFrame}; kwargs...)

Multiple miniature topoplots in regular distances.

Arguments

f::Union{GridPosition, GridLayout, GridLayoutBase.GridSubposition, Figure}
Figure, GridLayout, GridPosition, or GridLayoutBase.GridSubposition to draw the plot.

data::Union{<:Observable{<:DataFrame},DataFrame}
DataFrame with data or Observable DataFrame.
Requires a time column by default, but can be overridden by specifying mapping=(; x=:my_column) with any continuous or categorical column.

Keyword arguments (kwargs)

bin_width::Real = nothing
Number specifing the width of bin of continuous x-value in its units.

bin_num::Real = nothing
Number of topoplots.
Either bin_width, or bin_num should be specified. Error if they are both specified
If mapping.col or mapping.row are categorical bin_width and bin_num stay as nothing.

combinefun::Function = mean
Specify how the samples within bin_width are summarised.
Example functions: mean, median, std.

rasterize_heatmaps::Bool = true
Force rasterization of the plot heatmap when saving in svg format.
Except for the interpolated heatmap, all lines/points are vectors.
This is typically what you want, otherwise you get ~128x128 vectors per topoplot, which makes everything very slow.

col_labels::Bool, row_labels::Bool = true
Shows column and row labels for categorical values.

labels::Vector{String} = nothing
Show labels for each electrode.

positions::Vector{Point{2, Float32}} = nothing
Specify channel positions. Requires the list of x and y positions for all unique electrode.

interactive_scatter = nothing
Enable interactive mode.
If you create obs_tuple = Observable((0, 0, 0)) and pass it into interactive_scatter you can update the observable tuple with the indices of the clicked topoplot markers.
(0, 0, 0) corresponds to the (row of topoplot layout, column of topoplot layout, electrode).

topoplot_axes::NamedTuple = (;)
Here you can flexibly change configurations of topoplots.
To see all options just type ?Axis in REPL.

mapping = (; col = :time, row = nothing, layout = nothing)
mapping.col - specify x-value, can be any continuous or categorical variable.
mapping.row - specify y-value, can be any continuous or categorical variable (not implemented yet).
mapping.layout - arranges topoplots by rows when equals :time.

visual.colorrange::2-element Vector{Int64}, colorbar.colorrange::2-element Vector{Int64}
First is resposnible for colorrange in topoplots, second - in colorbars. Ideally they should be the same.

Shared plot configuration options

The shared plot options can be used as follows: type = (; key = value, ...)).
For example, plot_x(...; layout = (; show_legend = true, legend_position = :right)).
Multiple defaults will be cycled until match.

Placing ; is important!

figure = NamedTuple() - use kwargs... of Makie.Figure

axis = (xlabel = "Time windows [s]", aspect = Makie.DataAspect(), title = "", titlesize = 16, titlefont = :bold, ylabel = "", ylabelpadding = 25, xlabelpadding = 25, xpanlock = true, ypanlock = true, xzoomlock = true, yzoomlock = true, xrectzoom = false, yrectzoom = false) - use kwargs... of Makie.Axis

layout = (show_legend = true, use_colorbar = true, hidespines = (), hidedecorations = Dict{Symbol, Bool}(:label => 0))

mapping = (x = nothing, y = (:estimate, :yhat, :y), positions = (:pos, :positions, :position, nothing), labels = (:labels, :label, :sensor, nothing), col = (:time,), row = (nothing,))

visual = (colormap = Makie.Reverse{Symbol}(:RdBu), contours = (color = :white, linewidth = 2), enlarge = 1, label_scatter = false, label_text = false, bounding_geometry = GeometryBasics.Circle, levels = nothing) - use kwargs... of Topoplot.eeg_topoplot

legend = (orientation = :vertical, tellwidth = true, tellheight = false, halign = :right, valign = :center) - use kwargs... of Makie.Legend

colorbar = (vertical = true, tellwidth = true, tellheight = false, flipaxis = true, labelrotation = -1.5707963267948966, label = "Voltage [µV]", colorrange = nothing) - use kwargs... of Makie.Colorbar

Return Value: Figure displaying the Topoplot series.

UnfoldMakie.rel_to_abs_bbox — Method

rel_to_abs_bbox(org, rel)

Takes a rectangle org consiting of coordinates of origins and applies the relative transformation tuple rel.

Return Value: Makie.BBox.

UnfoldMakie.to_positions — Method

to_positions(x, y, z; sphere = [0, 0, 0.])
to_positions(pos::AbstractMatrix; sphere = [0, 0, 0.])

Projects 3D electrode positions to a 2D layout. Reimplementation of the MNE algorithm.

Assumes size(pos) = (3, nChannels) when input is AbstractMatrix.

Tip: You can get positions directly from an MNE object after loading PyMNE and enabling the UnfoldMakie PyMNE extension.

Return Value: Vector{Point2{Float64}}.

UnfoldMakie.update_axis! — Method

update_axis(support_axis::NamedTuple; kwargs...)

Update values of NamedTuple{key = value}
Used for supportive axes to make users be able to flexibly change them.

Internally, we use a PlotConfig struct to keep track of common plotting options, so that all functions have a similar API.

UnfoldMakie.PlotConfig — Type

PlotConfig(<plotname>)

Contains several different fields that can modify various aspects of the plot.

UnfoldMakie.PCPTicks — Type

PCPTicks

Used to inject extrema ticks and round them if necessary.